move hir db
