+++ crates/ra_analysis/src/hir/mod.rs
@@ -7,25 +7,18 @@
 
 pub(crate) mod function;
 pub(crate) mod module;
+pub(crate) mod db;
 mod path;
 
-use std::sync::Arc;
-
 use ra_syntax::{
-    ast::{self, FnDefNode, AstNode},
-    TextRange, SyntaxNode,
+    ast::{self, AstNode},
+    TextRange,
 };
 
 use crate::{
-    FileId,
-    db::SyntaxDatabase,
-    hir::function::{resolve_local_name, FnId, FnScopes},
-    hir::module::{
-        ModuleId, ModuleTree, ModuleSource,
-        nameres::{ItemMap, InputModuleItems, FileItems}
-    },
-    input::SourceRootId,
-    loc2id::{IdDatabase, DefId, DefLoc},
+    hir::db::HirDatabase,
+    hir::function::{resolve_local_name, FnScopes},
+    loc2id::{DefId, DefLoc},
     syntax_ptr::LocalSyntaxPtr,
     Cancelable,
 };
@@ -35,50 +28,6 @@ pub(crate) use self::{
     module::{ModuleDescriptor, nameres::FileItemId},
     function::FunctionDescriptor,
 };
-
-salsa::query_group! {
-pub(crate) trait HirDatabase: SyntaxDatabase + IdDatabase {
-        fn fn_scopes(fn_id: FnId) -> Arc<FnScopes> {
-            type FnScopesQuery;
-            use fn function::imp::fn_scopes;
-        }
-
-        fn _file_items(file_id: FileId) -> Arc<FileItems> {
-            type FileItemsQuery;
-            storage dependencies;
-            use fn module::nameres::file_items;
-        }
-
-        fn _file_item(file_id: FileId, file_item_id: FileItemId) -> SyntaxNode {
-            type FileItemQuery;
-            storage dependencies;
-            use fn module::nameres::file_item;
-        }
-
-        fn _input_module_items(source_root_id: SourceRootId, module_id: ModuleId) -> Cancelable<Arc<InputModuleItems>> {
-            type InputModuleItemsQuery;
-            use fn module::nameres::input_module_items;
-        }
-        fn _item_map(source_root_id: SourceRootId) -> Cancelable<Arc<ItemMap>> {
-            type ItemMapQuery;
-            use fn module::nameres::item_map;
-        }
-        fn _module_tree(source_root_id: SourceRootId) -> Cancelable<Arc<ModuleTree>> {
-            type ModuleTreeQuery;
-            use fn module::imp::module_tree;
-        }
-        fn _fn_syntax(fn_id: FnId) -> FnDefNode {
-            type FnSyntaxQuery;
-            // Don't retain syntax trees in memory
-            storage dependencies;
-            use fn function::imp::fn_syntax;
-        }
-        fn _submodules(source: ModuleSource) -> Cancelable<Arc<Vec<module::imp::Submodule>>> {
-            type SubmodulesQuery;
-            use fn module::imp::submodules;
-        }
-    }
-}
 
 pub(crate) enum Def {
     Module(ModuleDescriptor),
